Implement Client#stw
